+++ flodl/src/distributed/mod.rs
@@ -66,6 +66,3 @@ pub use ddp::{Ddp, DdpConfig, HasGraph, Trainer};
 pub use el_che::{ElChe, Phase};
 pub use lr_event_meta::{LrEventMeta, LrEventMetaConfig, MetaAction};
 pub use ddp_run::{ApplyPolicy, DdpHandle, DdpBuilder, DdpRunConfig, AverageBackend, TrainedState, EpochMetrics, MetricsFn, record_scalar, drain_scalars, GpuWorker};
-// Deprecated aliases
-#[allow(deprecated)]
-pub use ddp_run::{AsyncDdp, AsyncDdpBuilder, AsyncDdpConfig};